Adding graphs to base DAShboard
Added three of the graphs. Had to create new JS file to hold code. has a function that should input the new data into the graphs, needs to be tested still. Will also need to make a listening function to resize the graphs when the window is resized, at the moment it needs to be refreshed when the window size is changed.